requirements.txt
# Core web framework
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-APScheduler==1.13.0
python-dotenv==1.0.0
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.17
alembic==1.11.1

# Computer vision and pose detection
opencv-python==4.8.0.74
mediapipe==0.10.2
numpy==1.24.3

# AI/ML 
google-generativeai==0.3.1
protobuf==4.23.4

# Utilities
requests==2.31.0
jsonschema==4.17.3
Pillow==10.0.0